+++ sys/src/9/ppc/mmu.c
@@ -262,7 +262,7 @@ putmmu(ulong va, ulong pa, Page *pg)
 }
 
 void
-checkmmu(ulong, ulong)
+checkmmu(uintptr, uintptr)
 {
 }
